refactor(core): use relative imports within module boundary
